Develop

+++ controllers/ProjectContoller.js
@@ -1,15 +1,35 @@
 const ProjectModel = require("../models/ProjectModel");
+const cloudinary = require("../config/cloudinary");
+const { uploadImageToCloudinary } = require("../services/imageService");
 const { getColleaguesNames } = require("../services/projectService");
+const { notifyAllSubscribers } = require("../services/emailService");
 
 const addProject = async (req, res) => {
   try {
     const projectData = req.body;
+    const images = req.files;
 
     const newProject = new ProjectModel({ ...projectData });
+
+    if (images && images.length > 0) {
+      const uploadedImages = await Promise.all(
+        images.map((image) => uploadImageToCloudinary(image.buffer))
+      );
+
+      newProject.images = uploadedImages.map((newImage) => ({
+        url: newImage.url,
+        id: newImage.public_id,
+      }));
+    }
 
     await newProject.save();
 
     const { __v, ...project } = newProject.toObject();
+
+    const emailSubject = "Team Umeå";
+    const emailText = `Vi har publicerat ett nytt projekt på vår sida! Kolla in vårt senaste projekt: "${projectData.project}".`;
+
+    await notifyAllSubscribers(emailSubject, emailText);
 
     res.status(201).json({ message: "Projekt har lagts till", project, success: true });
   } catch (error) {
@@ -20,15 +40,64 @@ const addProject = async (req, res) => {
 
 const editProject = async (req, res) => {
   try {
-    const projectData = req.body;
+    const { _id, ...projectData } = req.body;
+    const images = req.files;
+    const existingProject = await ProjectModel.findById(_id);
 
-    await ProjectModel.updateOne(
-      { _id: projectData._id },
-      { $set: { ...projectData } },
-      { runValidators: true }
+    if (!existingProject) {
+      return res.status(404).json({ message: "Projekt kunde inte hittas", success: false });
+    }
+
+    const updatedProject = await ProjectModel.findOneAndUpdate(
+      { _id },
+      { $set: { ...projectData, images: existingProject.images } },
+      { new: true, runValidators: true }
     );
 
-    res.status(200).json({ message: "Projekt har uppdaterats", success: true });
+    if (!updatedProject) {
+      return res.status(404).json({ message: "Projekt kunde inte uppdateras", success: false });
+    }
+
+    const imageUrls =
+      typeof projectData.images === "string"
+        ? [projectData.images]
+        : projectData.images
+        ? projectData.images
+        : [];
+
+    const imagesToDestroy = existingProject.images.filter((img) => !imageUrls.includes(img.url));
+    const imagesToKeep = existingProject.images.filter((img) => imageUrls.includes(img.url));
+
+    if (images && images.length > 0) {
+      if (existingProject.images && existingProject.images.length > 0) {
+        await Promise.all(imagesToDestroy.map((image) => cloudinary.uploader.destroy(image.id)));
+      }
+
+      const uploadedImages = await Promise.all(
+        images.map((image) => uploadImageToCloudinary(image.buffer))
+      );
+
+      const mappedUploadedImages = uploadedImages.map((newImage) => ({
+        url: newImage.url,
+        id: newImage.public_id,
+      }));
+
+      updatedProject.images = [...imagesToKeep, ...mappedUploadedImages];
+
+      await updatedProject.save();
+    } else {
+      const uploadedImages = projectData?.images || [];
+      const existingImages = updatedProject.images;
+      const imagesToKeep = existingImages.filter((img) => uploadedImages.includes(img.url));
+      updatedProject.images = imagesToKeep;
+      await updatedProject.save();
+    }
+
+    const { _id: projectId, __v, ...projectDetails } = updatedProject.toObject();
+
+    res
+      .status(200)
+      .json({ message: "Projekt har uppdaterats", project: projectDetails, success: true });
   } catch (error) {
     console.error(error);
     res.status(500).json({ message: "Serverfel", success: false });
@@ -42,7 +111,8 @@ const getProjects = async (_, res) => {
     const projectsWithColleaguesNames = await Promise.all(
       projects.map(async (project) => {
         const colleaguesNames = await getColleaguesNames(project.colleagues);
-        return { ...project, colleagues: colleaguesNames };
+        const images = project?.images?.map((img) => img.url);
+        return { ...project, colleagues: colleaguesNames, images };
       })
     );
 
@@ -64,10 +134,12 @@ const getProjectById = async (req, res) => {
     }
 
     const colleaguesNames = await getColleaguesNames(project.colleagues);
+    const images = project.images.map((img) => img.url);
 
     const projectWithColleaguesNames = {
       ...project,
       colleagues: colleaguesNames,
+      images,
     };
 
     res.status(201).json({ project: projectWithColleaguesNames, success: true });
@@ -81,6 +153,18 @@ const deleteProject = async (req, res) => {
   const { projectid: projectID } = req.query;
 
   try {
+    const existingProject = await ProjectModel.findById(projectID);
+
+    if (!existingProject) {
+      return res.status(404).json({ message: "Projekt kunde inte hittas", success: false });
+    }
+
+    if (existingProject.images && existingProject.images.length > 0) {
+      await Promise.all(
+        existingProject.images.map((image) => cloudinary.uploader.destroy(image.id))
+      );
+    }
+
     await ProjectModel.deleteOne({ _id: projectID });
 
     res.status(204).json({ message: "Projekt har raderats", success: true });
